central/central-api/requirements.txt
Flask==2.0.1
Flask-SocketIO==5.1.1
psycopg2-binary==2.9.1  # PostgreSQL database driver